% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme_color_function.R
\name{gimme_color_codes}
\alias{gimme_color_codes}
\title{RLadies palette function}
\usage{
gimme_color_codes(wantAll = TRUE)
}
\arguments{
\item{wantALL}{Do you want all the colors available in this palette? The default is TRUE}
}
\description{
This function will print out the hex codes for the RLadies palette
}
\examples{
gimme_color_codes()

}
\keyword{RLadies}
\keyword{colors}
