% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rladiesgnv-package.R
\docType{package}
\name{rladiesgnv-package}
\alias{rladiesgnv}
\alias{rladiesgnv-package}
\title{rladiesgnv: Gets you the pretty colors from the R Ladies palette}
\description{
This is a simple package with one function. It was created as part of a tutorial for the R Ladies Gainesville meeting on April 29, 2019.
}
\author{
\strong{Maintainer}: Javiera Rudolph \email{javirudolph@outlook.com}

}
\keyword{internal}
